+++ backend/models/agent_risk_profile.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from backend.database import Base
@@ -11,7 +11,7 @@ class AgentRiskProfile(Base):
     
     credit_score = Column(Integer, default=0) # 0-100
     risk_level = Column(String, default="unknown") # e.g., "Very Safe", "Safe", "Medium", "Risky", "Dangerous"
-    recommended_limit = Column(Float, default=0.0) # Max eligible loan amount
+    recommended_limit = Column(Numeric(20, 2), default=0.0) # Max eligible loan amount
     last_calculated = Column(DateTime(timezone=True), server_default=func.now())
 
     agent = relationship("Agent", back_populates="risk_profile")
